requirements.txt
fastapi==0.95.2
uvicorn[standard]
requests
pandas
youtube_transcript_api
python-dotenv
openai
langchain
tiktoken
faiss-cpu
pinecone-client
fastapi_async_langchain
bs4
